refactor(users): migrate Users API to module structure
Move the Users service from app/controllers/api/users.php to
src/Appwrite/Platform/Modules/Users, one action class per endpoint.

HTTP directories mirror the URL path, so scrypt-modified lives under
Users/Scrypt/Modified/Create.php and phone verification under
Users/Verification/Phone/Update.php. The two session deletes are split
into Sessions/Delete.php for the item route and Sessions/Bulk/Delete.php
for the collection route, matching the convention used by TablesDB rows.

The global createUser() helper becomes a protected method on
Modules/Users/Base.php, which the eight hash-variant create actions
extend.

Paths, methods, scopes, events, audit labels and SDK Method definitions
are unchanged, so the public API and generated SDKs stay identical.
