+++ app/config/services.php
@@ -174,7 +174,7 @@ return [
         'name' => 'Users',
         'subtitle' => 'The Users service allows you to manage your project users.',
         'description' => '/docs/services/users.md',
-        'controller' => 'api/users.php',
+        'controller' => '', // Uses modules
         'sdk' => true,
         'docs' => true,
         'docsUrl' => 'https://appwrite.io/docs/server/users',
