+++ src/Appwrite/Platform/Appwrite.php
@@ -21,6 +21,7 @@ use Appwrite\Platform\Modules\Sites;
 use Appwrite\Platform\Modules\Storage;
 use Appwrite\Platform\Modules\Teams;
 use Appwrite\Platform\Modules\Tokens;
+use Appwrite\Platform\Modules\Users;
 use Appwrite\Platform\Modules\VCS;
 use Appwrite\Platform\Modules\Webhooks;
 use Utopia\Platform\Platform;
@@ -43,6 +44,7 @@ class Appwrite extends Platform
         $this->addModule(new Proxy\Module());
         $this->addModule(new Teams\Module());
         $this->addModule(new Tokens\Module());
+        $this->addModule(new Users\Module());
         $this->addModule(new Storage\Module());
         $this->addModule(new VCS\Module());
         $this->addModule(new Webhooks\Module());
