+++ src/Appwrite/Platform/Modules/Users/Http/Users/Create.php
@@ -1,0 +1,82 @@
+<?php
+
+namespace Appwrite\Platform\Modules\Users\Http\Users;
+
+use Appwrite\Auth\Validator\PasswordDictionary;
+use Appwrite\Auth\Validator\PasswordStrength;
+use Appwrite\Auth\Validator\Phone;
+use Appwrite\Hooks\Hooks;
+use Appwrite\Platform\Action;
+use Appwrite\Platform\Modules\Users\Base;
+use Appwrite\SDK\AuthType;
+use Appwrite\SDK\Method;
+use Appwrite\SDK\Response as SDKResponse;
+use Appwrite\SDK\Specification\Validator\PasswordFormat;
+use Appwrite\Utopia\Database\Validator\CustomId;
+use Appwrite\Utopia\Response;
+use Utopia\Auth\Hashes\Plaintext;
+use Utopia\Database\Database;
+use Utopia\Database\Document;
+use Utopia\Emails\Validator\Email as EmailValidator;
+use Utopia\Platform\Scope\HTTP;
+use Utopia\Validator;
+use Utopia\Validator\AllOf;
+use Utopia\Validator\Nullable;
+use Utopia\Validator\Text;
+
+class Create extends Base
+{
+    use HTTP;
+
+    public static function getName(): string
+    {
+        return 'createUser';
+    }
+
+    public function __construct()
+    {
+        $this
+            ->setHttpMethod(Action::HTTP_REQUEST_METHOD_POST)
+            ->setHttpPath('/v1/users')
+            ->desc('Create user')
+            ->groups(['api', 'users'])
+            ->label('scope', 'users.write')
+            ->label('audits.event', 'user.create')
+            ->label('audits.resource', 'user/{response.$id}')
+            ->label('sdk', new Method(
+                namespace: 'users',
+                group: 'users',
+                name: 'create',
+                description: '/docs/references/users/create-user.md',
+                auth: [AuthType::ADMIN, AuthType::KEY],
+                responses: [
+                    new SDKResponse(
+                        code: Response::STATUS_CODE_CREATED,
+                        model: Response::MODEL_USER,
+                    )
+                ]
+            ))
+            ->param('userId', '', fn (Database $dbForProject) => new CustomId(false, $dbForProject->getAdapter()->getMaxUIDLength()), 'User ID. Choose a custom ID or generate a random ID with `ID.unique()`. Valid chars are a-z, A-Z, 0-9, period, hyphen, and underscore. Can\'t start with a special char. Max length is 36 chars.', false, ['dbForProject'])
+            ->param('email', null, new Nullable(new EmailValidator()), 'User email.', true)
+            ->param('phone', null, new Nullable(new Phone()), 'Phone number. Format this number with a leading \'+\' and a country code, e.g., +16175551212.', true)
+            ->param('password', '', fn ($project, $passwordsDictionary) => new PasswordFormat(new AllOf([new PasswordStrength($project->getAttribute('auths', [])['passwordStrength'] ?? []), new PasswordDictionary($passwordsDictionary, enabled: $project->getAttribute('auths', [])['passwordDictionary'] ?? false)], Validator::TYPE_STRING)), 'Plain text user password. Must be at least 8 chars.', true, ['project', 'passwordsDictionary'])
+            ->param('name', '', new Text(128), 'User name. Max length: 128 chars.', true)
+            ->inject('response')
+            ->inject('project')
+            ->inject('dbForProject')
+            ->inject('hooks')
+            ->inject('plan')
+            ->callback($this->action(...));
+    }
+
+    public function action(string $userId, ?string $email, ?string $phone, ?string $password, ?string $name, Response $response, Document $project, Database $dbForProject, Hooks $hooks, array $plan): void
+    {
+        $plaintext = new Plaintext();
+
+        $user = $this->createUser($plaintext, $userId, $email, $password, $phone, $name, $project, $dbForProject, $hooks, $plan);
+
+        $response
+            ->setStatusCode(Response::STATUS_CODE_CREATED)
+            ->dynamic($user, Response::MODEL_USER);
+    }
+}
