+++ src/Appwrite/Platform/Modules/Users/Http/Users/MFA/RecoveryCodes/Create.php
@@ -1,0 +1,107 @@
+<?php
+
+namespace Appwrite\Platform\Modules\Users\Http\Users\MFA\RecoveryCodes;
+
+use Appwrite\Auth\MFA\Type;
+use Appwrite\Event\Event;
+use Appwrite\Extend\Exception;
+use Appwrite\Platform\Action;
+use Appwrite\SDK\AuthType;
+use Appwrite\SDK\Deprecated;
+use Appwrite\SDK\Method;
+use Appwrite\SDK\Response as SDKResponse;
+use Appwrite\Utopia\Response;
+use Utopia\Database\Database;
+use Utopia\Database\Document;
+use Utopia\Database\Validator\UID;
+use Utopia\Platform\Scope\HTTP;
+
+class Create extends Action
+{
+    use HTTP;
+
+    public static function getName(): string
+    {
+        return 'createUserMFARecoveryCodes';
+    }
+
+    public function __construct()
+    {
+        $this
+            ->setHttpMethod(Action::HTTP_REQUEST_METHOD_PATCH)
+            ->setHttpPath('/v1/users/:userId/mfa/recovery-codes')
+            ->desc('Create MFA recovery codes')
+            ->groups(['api', 'users'])
+            ->label('event', 'users.[userId].create.mfa.recovery-codes')
+            ->label('scope', 'users.write')
+            ->label('audits.event', 'user.update')
+            ->label('audits.resource', 'user/{response.$id}')
+            ->label('audits.userId', '{response.$id}')
+            ->label('usage.metric', 'users.{scope}.requests.update')
+            ->label('sdk', [
+                new Method(
+                    namespace: 'users',
+                    group: 'mfa',
+                    name: 'createMfaRecoveryCodes',
+                    description: '/docs/references/users/create-mfa-recovery-codes.md',
+                    auth: [AuthType::ADMIN, AuthType::KEY],
+                    responses: [
+                        new SDKResponse(
+                            code: Response::STATUS_CODE_CREATED,
+                            model: Response::MODEL_MFA_RECOVERY_CODES,
+                        )
+                    ],
+                    deprecated: new Deprecated(
+                        since: '1.8.0',
+                        replaceWith: 'users.createMFARecoveryCodes',
+                    ),
+                    public: false,
+                ),
+                new Method(
+                    namespace: 'users',
+                    group: 'mfa',
+                    name: 'createMFARecoveryCodes',
+                    description: '/docs/references/users/create-mfa-recovery-codes.md',
+                    auth: [AuthType::ADMIN, AuthType::KEY],
+                    responses: [
+                        new SDKResponse(
+                            code: Response::STATUS_CODE_CREATED,
+                            model: Response::MODEL_MFA_RECOVERY_CODES,
+                        )
+                    ]
+                )
+            ])
+            ->param('userId', '', fn (Database $dbForProject) => new UID($dbForProject->getAdapter()->getMaxUIDLength()), 'User ID.', false, ['dbForProject'])
+            ->inject('response')
+            ->inject('dbForProject')
+            ->inject('queueForEvents')
+            ->callback($this->action(...));
+    }
+
+    public function action(string $userId, Response $response, Database $dbForProject, Event $queueForEvents): void
+    {
+        $user = $dbForProject->getDocument('users', $userId);
+
+        if ($user->isEmpty()) {
+            throw new Exception(Exception::USER_NOT_FOUND);
+        }
+
+        $mfaRecoveryCodes = $user->getAttribute('mfaRecoveryCodes', []);
+
+        if (!empty($mfaRecoveryCodes)) {
+            throw new Exception(Exception::USER_RECOVERY_CODES_ALREADY_EXISTS);
+        }
+
+        $mfaRecoveryCodes = Type::generateBackupCodes();
+        $user->setAttribute('mfaRecoveryCodes', $mfaRecoveryCodes);
+        $dbForProject->updateDocument('users', $user->getId(), new Document(['mfaRecoveryCodes' => $mfaRecoveryCodes]));
+
+        $queueForEvents->setParam('userId', $user->getId());
+
+        $document = new Document([
+            'recoveryCodes' => $mfaRecoveryCodes
+        ]);
+
+        $response->dynamic($document, Response::MODEL_MFA_RECOVERY_CODES);
+    }
+}
