+++ src/Appwrite/Platform/Modules/Users/Http/Users/Sessions/Bulk/Delete.php
@@ -1,0 +1,83 @@
+<?php
+
+namespace Appwrite\Platform\Modules\Users\Http\Users\Sessions\Bulk;
+
+use Appwrite\Event\Event;
+use Appwrite\Extend\Exception;
+use Appwrite\Platform\Action;
+use Appwrite\SDK\AuthType;
+use Appwrite\SDK\ContentType;
+use Appwrite\SDK\Method;
+use Appwrite\SDK\Response as SDKResponse;
+use Appwrite\Utopia\Response;
+use Utopia\Database\Database;
+use Utopia\Database\Document;
+use Utopia\Database\Validator\UID;
+use Utopia\Platform\Scope\HTTP;
+
+class Delete extends Action
+{
+    use HTTP;
+
+    public static function getName(): string
+    {
+        return 'deleteUserSessions';
+    }
+
+    public function __construct()
+    {
+        $this
+            ->setHttpMethod(Action::HTTP_REQUEST_METHOD_DELETE)
+            ->setHttpPath('/v1/users/:userId/sessions')
+            ->desc('Delete user sessions')
+            ->groups(['api', 'users'])
+            ->label('event', 'users.[userId].sessions.delete')
+            ->label('scope', ['users.write', 'sessions.write'])
+            ->label('audits.event', 'session.delete')
+            ->label('audits.resource', 'user/{user.$id}')
+            ->label('sdk', new Method(
+                namespace: 'users',
+                group: 'sessions',
+                name: 'deleteSessions',
+                description: '/docs/references/users/delete-user-sessions.md',
+                auth: [AuthType::ADMIN, AuthType::KEY],
+                responses: [
+                    new SDKResponse(
+                        code: Response::STATUS_CODE_NOCONTENT,
+                        model: Response::MODEL_NONE,
+                    )
+                ],
+                contentType: ContentType::NONE
+            ))
+            ->param('userId', '', fn (Database $dbForProject) => new UID($dbForProject->getAdapter()->getMaxUIDLength()), 'User ID.', false, ['dbForProject'])
+            ->inject('response')
+            ->inject('dbForProject')
+            ->inject('queueForEvents')
+            ->callback($this->action(...));
+    }
+
+    public function action(string $userId, Response $response, Database $dbForProject, Event $queueForEvents): void
+    {
+        $user = $dbForProject->getDocument('users', $userId);
+
+        if ($user->isEmpty()) {
+            throw new Exception(Exception::USER_NOT_FOUND);
+        }
+
+        $sessions = $user->getAttribute('sessions', []);
+
+        foreach ($sessions as $key => $session) {
+            /** @var Document $session */
+            $dbForProject->deleteDocument('sessions', $session->getId());
+            //TODO: fix this
+        }
+
+        $dbForProject->purgeCachedDocument('users', $user->getId());
+
+        $queueForEvents
+            ->setParam('userId', $user->getId())
+            ->setPayload($response->output($user, Response::MODEL_USER));
+
+        $response->noContent();
+    }
+}
